test: refactor setupDummyOOBM into parent class to resolve DRY violation
